pyproject.toml
[project]
name = "tub-lifter"
version = "0.1.0"
description = "Web app to control a Thomson Electrak HD linear actuator via CAN bus"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "python-can>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tub_lifter"]

[project.scripts]
tub-lifter = "tub_lifter.app:main"
